Refactoring into more sensible folder structure
