+++ src/WordGame/WordHelpers/IWordHelper.cs
@@ -1,7 +1,8 @@
-﻿namespace Chat.WordGame.WebHelpers
+﻿namespace Chat.WordGame.WordHelpers
 {
     public interface IWordHelper
     {
+        bool StrippedSuffixDictionaryCheck(string word);
         bool CheckWordEndingExists(string word);
     }
 }
